--- a/src/coreclr/src/classlibnative/bcltype/system.h
+++ b/src/coreclr/src/classlibnative/bcltype/system.h
@@ -43,13 +43,9 @@ private:
 
 public:
     // Functions on the System.Environment class
-#ifndef TARGET_UNIX
-    static FCDECL1(VOID, GetSystemTimeWithLeapSecondsHandling, FullSystemTime *time);
-    static FCDECL2(FC_BOOL_RET, ValidateSystemTime, SYSTEMTIME *time, CLR_BOOL localTime);
-    static FCDECL2(FC_BOOL_RET, FileTimeToSystemTime, INT64 fileTime, FullSystemTime *time);
-    static FCDECL2(FC_BOOL_RET, SystemTimeToFileTime, SYSTEMTIME *time, INT64 *pFileTime);
-#endif // TARGET_UNIX
+#ifdef TARGET_UNIX
     static FCDECL0(INT64, __GetSystemTimeAsFileTime);
+#endif // TARGET_UNIX
     static FCDECL0(UINT32, GetTickCount);
     static FCDECL0(UINT64, GetTickCount64);