Resolves https://github.com/dotnet/runtime/issues/13091.

This PR is based on a series of conversations @tarekgh and I have had recently with regard to short-circuiting our leap second handling logic where possible. For background, Windows understands that days normally have 86,400 seconds, but this can be off by ±1 second due to positive or negative leap seconds. .NET's DateTime type requires each day to have _exactly_ 86,400 seconds. Previous .NET releases resolve this discrepancy through the following special-cased logic: "Is the current time 23:59:60.xxx? If so, I'll return a DateTime instance whose time component reads 23:59:59.999." Unfortunately, this check is not free, and it shows up in benchmarks when we access `DateTime.UtcNow`.

This PR relies on the following properties of leap seconds:

* leap seconds only ever affect the last day of the month, _and_
* leap seconds only ever kick in on the last second of the day, _and_
* any given month will never be adjusted beyond ±1 second due to leap seconds.

This optimization works by caching the OS-reported tick value of 00h00:00 on the first day of the month, the corresponding .NET tick value, and the count of how many ticks would elapse between that and 23h59:59 on the final day of the month. If the OS-reported tick value is between 00h00:00 first day (inclusive) and 23h59:59 final day (exclusive), we know a leap second could not have occurred within that time interval, so we bypass the normal leap second handling logic and perform the equivalent of `DateTime.AddTicks` against the baseline "start of month" DateTime value. If the cache is stale, we update the cache. If we're within the small window at the end of the month where a leap second _might_ kick in, we ignore the cache and fall down the slower path.

Another change included in this PR is to remove the unmanaged implementation of the Windows `DateTime` logic, leaving the entire implementation fully managed except for the p/invokes directly into the operating system. This also unifies the Mono and CoreCLR implementations. The non-Windows implementation continues to be split across the managed and unmanaged world.

### Perf results

|    Method |        Job | Toolchain |     Mean |    Error |   StdDev | Ratio |
|---------- |----------- |---------- |---------:|---------:|---------:|------:|
| GetUtcNow | Job-WLQNMC |    master | 64.61 ns | 0.431 ns | 0.403 ns |  1.00 |
| GetUtcNow | Job-GWWSKI |     proto | 28.45 ns | 0.167 ns | 0.148 ns |  0.44 |

### Todo before merge

- [ ] Run perf analysis showing that switching to a fully-managed implementation doesn't degrade perf
- [ ] Run Tarek's tool which manipulates the OS leap second data to exercise edge cases in this code
- [ ] Get extra eyes from the Windows crew to validate any assumptions we've made here
- [ ] Exercise edge cases manually inside a debugger
- [ ] Minor code cleanup to follow coreclr conventions