@GrabYourPitchforks 2020-11-17
Some extra perf numbers:

|    Method |        Job |                Toolchain |     Mean |    Error |   StdDev | Ratio |
|---------- |----------- |------------------------- |---------:|---------:|---------:|------:|
| GetUtcNow | Job-IPXYOZ |                   master | 64.96 ns | 0.405 ns | 0.379 ns |  1.00 |
| GetUtcNow | Job-NFLWIO |            proto_managed | 28.79 ns | 0.093 ns | 0.087 ns |  0.44 |
| GetUtcNow | Job-VXWZCK | proto_managed_suppressgc | 27.29 ns | 0.210 ns | 0.186 ns |  0.42 |
| GetUtcNow | Job-TGFBZW |              proto_mixed | 27.72 ns | 0.151 ns | 0.134 ns |  0.43 |
| GetUtcNow | Job-VYTQZD |   proto_mixed_suppressgc | 28.05 ns | 0.201 ns | 0.168 ns |  0.43 |

* **proto_managed** - the PR as it currently stands, with unmanaged callconv calli into the OS.
* **proto_managed_suppressgc** - as above, but using `delegate* unmanaged[StdCall, SuppressGCTransition]<...>`.
* **proto_mixed** - keeping CoreCLR's `GetSystemTimeAsFileTime` method around (as a FCall) and using it instead of calli into OS.
* **proto_mixed_suppressgc** - as above, but turning the method into a QCall instead of an FCall and using `[SuppressGCTransition]`.

I don't know why **proto_mixed_suppressgc** clocks in at slightly slower than **proto_mixed**. The difference isn't enough to worry about, but I would've expected after many runs for them to be within each other's stddev. Maybe a bad conversion on my part?