@GrabYourPitchforks 2020-11-17
@adamsitnik @tarekgh In the latest iteration, I split the cache lookup [into its own helper method](https://github.com/dotnet/runtime/pull/44771/commits/35d60b87801ec3af8e2acada7ebb5c2289441f6b#diff-d82d22c9b803839953afdb8156738425f56e102241c1d4ad9f19439858d7ace5R98), separate from the OS p/invoke. This way, we can write a unit test which doesn't muck about with the calli target, which I was afraid could introduce stability problems (especially once `[SuppressGCTransition]` comes online). Now, what we can do from within the unit test is to bypass the p/invoke entirely, using private reflection to pass a known FILETIME value into the helper method and observing that the results make sense.

Such unit test isn't written yet; this is just setting up for that.

Do you think this approach satisfies the testing concerns you had?