@GrabYourPitchforks 2020-11-17
Here's the x64 codegen as of the latest iteration.

```asm
; System_Private_CoreLib!System.DateTime.get_UtcNow()
; lots of prolog due to the unmanaged calli
00007ff9`5105b5a0 55              push    rbp
00007ff9`5105b5a1 4157            push    r15
00007ff9`5105b5a3 4156            push    r14
00007ff9`5105b5a5 4155            push    r13
00007ff9`5105b5a7 4154            push    r12
00007ff9`5105b5a9 57              push    rdi
00007ff9`5105b5aa 56              push    rsi
00007ff9`5105b5ab 53              push    rbx
00007ff9`5105b5ac 4883ec78        sub     rsp,78h
00007ff9`5105b5b0 488dac24b0000000 lea     rbp,[rsp+0B0h]
00007ff9`5105b5b8 488d4d80        lea     rcx,[rbp-80h]
00007ff9`5105b5bc 498bd2          mov     rdx,r10
00007ff9`5105b5bf e81ce9885f      call    CoreCLR!JIT_InitPInvokeFrame (00007ff9`b08e9ee0)
00007ff9`5105b5c4 488bf0          mov     rsi,rax
00007ff9`5105b5c7 488bcc          mov     rcx,rsp
00007ff9`5105b5ca 48894da0        mov     qword ptr [rbp-60h],rcx
00007ff9`5105b5ce 488bcd          mov     rcx,rbp
00007ff9`5105b5d1 48894db0        mov     qword ptr [rbp-50h],rcx

; read fnptr from static field and invoke it
00007ff9`5105b5d5 48b9d808db50f97f0000 mov rcx,7FF950DB08D8h
00007ff9`5105b5df 488b01          mov     rax,qword ptr [rcx]
00007ff9`5105b5e2 488d4dc0        lea     rcx,[rbp-40h]
00007ff9`5105b5e6 488d1512000000  lea     rdx,[System_Private_CoreLib!System.DateTime.get_UtcNow()+0xffffffff`a70edd8f (00007ff9`5105b5ff)]
00007ff9`5105b5ed 488955a8        mov     qword ptr [rbp-58h],rdx
00007ff9`5105b5f1 488d5580        lea     rdx,[rbp-80h]
00007ff9`5105b5f5 48895610        mov     qword ptr [rsi+10h],rdx
00007ff9`5105b5f9 c6460c00        mov     byte ptr [rsi+0Ch],0
00007ff9`5105b5fd ffd0            call    rax
00007ff9`5105b5ff c6460c01        mov     byte ptr [rsi+0Ch],1
00007ff9`5105b603 48ba5c12cdb0f97f0000 mov rdx,offset CoreCLR!g_TrapReturningThreads (00007ff9`b0cd125c)
00007ff9`5105b60d 833a00          cmp     dword ptr [rdx],0
00007ff9`5105b610 740c            je      System_Private_CoreLib!System.DateTime.get_UtcNow()+0xffffffff`a70eddae (00007ff9`5105b61e)
00007ff9`5105b612 48b9f833ccb0f97f0000 mov rcx,offset CoreCLR!hlpDynamicFuncTable+0xa8 (00007ff9`b0cc33f8)
00007ff9`5105b61c ff11            call    qword ptr [rcx]
00007ff9`5105b61e 488b4d88        mov     rcx,qword ptr [rbp-78h]
00007ff9`5105b622 48894e10        mov     qword ptr [rsi+10h],rcx
00007ff9`5105b626 488b4dc0        mov     rcx,qword ptr [rbp-40h]

; read s_cache and make sure we're still in bounds
00007ff9`5105b62a 48b81812d31bec010000 mov rax,1EC1BD31218h
00007ff9`5105b634 488b00          mov     rax,qword ptr [rax]
00007ff9`5105b637 488bd1          mov     rdx,rcx
00007ff9`5105b63a 482b5008        sub     rdx,qword ptr [rax+8]
00007ff9`5105b63e 483b5018        cmp     rdx,qword ptr [rax+18h]
00007ff9`5105b642 7309            jae     System_Private_CoreLib!System.DateTime.get_UtcNow()+0xffffffff`a70edddd (00007ff9`5105b64d)
00007ff9`5105b644 48035010        add     rdx,qword ptr [rax+10h]
00007ff9`5105b648 488bc2          mov     rax,rdx
00007ff9`5105b64b eb05            jmp     System_Private_CoreLib!System.DateTime.get_UtcNow()+0xffffffff`a70edde2 (00007ff9`5105b652)

; call Fallback
00007ff9`5105b64d e8965adaff      call    CLRStub[MethodDescPrestub]@7ff950e010e8 (00007ff9`50e010e8)

; epilog
00007ff9`5105b652 90              nop
00007ff9`5105b653 488d65c8        lea     rsp,[rbp-38h]
00007ff9`5105b657 5b              pop     rbx
00007ff9`5105b658 5e              pop     rsi
00007ff9`5105b659 5f              pop     rdi
00007ff9`5105b65a 415c            pop     r12
00007ff9`5105b65c 415d            pop     r13
00007ff9`5105b65e 415e            pop     r14
00007ff9`5105b660 415f            pop     r15
00007ff9`5105b662 5d              pop     rbp
00007ff9`5105b663 c3              ret
```