@tarekgh 2020-11-17
> Do you think this approach satisfies the testing concerns you had?

I would say the best to have just a simple test that get the time from .NET and from the OS and then compare them. The difference should not exceed some threshold like 100 ms or so. 

@GrabYourPitchforks I have added the comment https://github.com/dotnet/runtime/pull/44771#discussion_r525422755 I hope it is not lost in the discussions here :-)