@adamsitnik 2020-11-18
> Do you think this approach satisfies the testing concerns you had?

@GrabYourPitchforks yes, it does! Thanks for working on this.

> I would say the best to have just a simple test that get the time from .NET and from the OS and then compare them. The difference should not exceed some threshold like 100 ms or so.

@tarekgh but how such a test would be exercising the edge cases?