@tarekgh 2020-11-18
> but how such a test would be exercising the edge cases?

This test is going to run with every PR we submit. for sure we'll hit the edge case live :-) we already hit such things with bad written tests when the daylight saving change and we see the failures.