@adamsitnik 2020-11-20
> This test is going to run with every PR we submit. for sure we'll hit the edge case live

I see. We should most probably have both:

- mocked tests for edge cases we are planning for and know of. this allows us for testing them before bad thigns happen
- .NET vs OS time test to verify the implementation and possibly search for edge cases that we were not aware of